Add comments and a version in the no-packing and NEON cases.
