+++ field/src/packed/mod.rs
@@ -10,3 +10,8 @@ pub use x86_64_avx::*;
 mod aarch64_neon;
 #[cfg(all(target_arch = "aarch64", target_feature = "neon"))]
 pub use aarch64_neon::*;
+
+#[cfg(not(any(all(target_arch = "aarch64", target_feature = "neon"), all(target_arch = "x86_64", target_feature = "avx2",))))]
+mod no_packing;
+#[cfg(not(any(all(target_arch = "aarch64", target_feature = "neon"), all(target_arch = "x86_64", target_feature = "avx2",))))]
+pub use no_packing::*;
