+++ field/src/packed/x86_64_avx.rs
@@ -95,6 +95,16 @@ pub fn mm512_mod_add(a: __m512i, b: __m512i, p: __m512i) -> __m512i {
     }
 }
 
+/// Add two arrays of integers modulo `P` using packings.
+///
+/// Assumes that `P` is less than `2^31` and `a + b <= 2P` for all array pairs `a, b`.
+/// If the inputs are not in this range, the result may be incorrect.
+/// The result will be in the range `[0, P]` and equal to `(a + b) mod P`.
+/// It will be equal to `P` if and only if `a + b = 2P` so provided `a + b < 2P`
+/// the result is guaranteed to be less than `P`.
+/// 
+/// Scalar add is assumed to be a function which implements `a + b % P` with the 
+/// same specifications as above.
 #[inline(always)]
 pub fn packed_mod_add<const WIDTH: usize>(
     a: &[u32; WIDTH],
