+++ monty-31/src/aarch64_neon/packing.rs
@@ -609,52 +609,3 @@ pub(crate) fn octic_mul_packed<FP, const WIDTH: usize>(
     res[..4].copy_from_slice(&dot_0);
     res[4..].copy_from_slice(&dot_1);
 }
-
-#[inline(always)]
-pub(crate) fn packed_add<FP, const WIDTH: usize>(
-    a: &[MontyField31<FP>; WIDTH],
-    b: &[MontyField31<FP>; WIDTH],
-) -> [MontyField31<FP>; WIDTH]
-where
-    FP: FieldParameters + BinomialExtensionData<WIDTH>,
-{
-    let mut res = [MontyField31::ZERO; WIDTH];
-    match WIDTH {
-        1 => res[0] = a[0] + b[0],
-        4 => {
-            // Perfectly fits into a PackedMontyField31Neon element.
-            let packed_a = PackedMontyField31Neon([a[0], a[1], a[2], a[3]]);
-            let packed_b = PackedMontyField31Neon([b[0], b[1], b[2], b[3]]);
-            let out = packed_a + packed_b;
-
-            res.copy_from_slice(&out.0);
-        }
-        5 => {
-            // We fit what we can into a PackedMontyField31Neon element.
-            // The final add is done using a scalar addition.
-            let packed_a = PackedMontyField31Neon([a[0], a[1], a[2], a[3]]);
-            let packed_b = PackedMontyField31Neon([b[0], b[1], b[2], b[3]]);
-            let out = packed_a + packed_b;
-
-            res.copy_from_slice(&out.0);
-            res[4] = a[4] + b[4];
-
-            res[..4].copy_from_slice(&out.0);
-        }
-        8 => {
-            // This perfectly fits into two PackedMontyField31Neon elements.
-            let packed_a_lo = PackedMontyField31Neon([a[0], a[1], a[2], a[3]]);
-            let packed_b_lo = PackedMontyField31Neon([b[0], b[1], b[2], b[3]]);
-            let out_lo = packed_a_lo + packed_b_lo;
-
-            let packed_a_hi = PackedMontyField31Neon([a[4], a[5], a[6], a[7]]);
-            let packed_b_hi = PackedMontyField31Neon([b[4], b[5], b[6], b[7]]);
-            let out_hi = packed_a_hi + packed_b_hi;
-
-            res[..4].copy_from_slice(&out_lo.0);
-            res[4..].copy_from_slice(&out_hi.0);
-        }
-        _ => panic!("Currently unsupported width for packed addition"),
-    }
-    res
-}
